+++ basketapp/admin.py
@@ -1,5 +1,4 @@
 from django.contrib import admin
-
 from basketapp.models import Basket
 
 admin.site.register(Basket)
