+++ geekshop/settings.py
@@ -41,7 +41,7 @@ INSTALLED_APPS = [
 
     'mainapp',
     'authapp',
-#    'basketapp',
+    'basketapp',
 ]
 
 AUTH_USER_MODEL = 'authapp.ShopUser'
