+++ geekshop/urls.py
@@ -21,11 +21,11 @@ from django.urls import path, include
 from geekshop.views import index, contacts
 
 urlpatterns = [
+    path('', index, name='index'),
     path('admin/', admin.site.urls),
     path('products/', include('mainapp.urls', namespace='products'), name='products'),
     path('auth/', include('authapp.urls', namespace='auth'), name='auth'),
-#    path('basket/', include('basketapp.urls', namespace='basket'), name='basket'),
-    path('', index, name='index'),
+    path('basket/', include('basketapp.urls', namespace='basket'), name='basket'),
     path('contacts/', contacts, name='contacts')
 ]
