--- mainapp/templates/mainapp/includes/inc_menu.html
@@ -1,1 +1,0 @@
-//todo
