+++ mainapp/urls.py
@@ -8,5 +8,5 @@ urlpatterns = [
     path('', products, name='index'),
     # path('category/<int:pk>/', mainapp.products, name='category'),
     path('category/<int:pk>/', products, name='category'),
-    path('product/<int:pk>/', product, name='product'),
+    # path('product/<int:pk>/', product, name='product'),
 ]
