+++ mainapp/views.py
@@ -2,15 +2,15 @@ import random
 
 from django.shortcuts import render, get_object_or_404
 
-#from basketapp.models import Basket
+from basketapp.models import Basket
 from mainapp.models import Product, ProductCategory
 
 
-# def get_basket(user):
-#     if user.is_authenticated:
-#         return Basket.objects.filter(user=user)
-#     else:
-#         return []
+def get_basket(user):
+    if user.is_authenticated:
+        return Basket.objects.filter(user=user)
+    else:
+        return []
 
 
 def get_same_products(hot_product):
@@ -28,7 +28,7 @@ def products(request, pk=None):
     title = 'продукты'
     links_menu = ProductCategory.objects.all()
     same_products = Product.objects.all()[:4]
-    # basket = get_basket(request.user)
+    basket = get_basket(request.user)
 
     hot_product = get_hot_product()
     same_products = get_same_products(hot_product)
@@ -47,7 +47,7 @@ def products(request, pk=None):
             'category': category,
             'related_products': same_products,
             'products': products,
-        #    'basket': basket,
+            'basket': basket,
             'hot_product': hot_product,
         }
         return render(request, 'mainapp/products.html', context)
@@ -61,7 +61,7 @@ def products(request, pk=None):
         'related_products': same_products,
         'hot_product': hot_product,
         'products': products,
-#        'basket': basket,
+        'basket': basket,
     }
     return render(request, 'mainapp/products.html', context)
 
@@ -78,7 +78,7 @@ def product(request, pk):
         'title': title,
         'links_menu': links_menu,
         'related_products': same_products,
-    #    'basket': basket,
+        'basket': basket,
         'product': product,
     }
     return render(request, 'mainapp/product.html', context)
